+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/DeveloperDecertificationResponse.java
@@ -1,0 +1,21 @@
+package gov.healthit.chpl.domain;
+
+import java.util.List;
+
+import gov.healthit.chpl.dto.DeveloperDecertifiedDTO;
+
+public class DeveloperDecertificationResponse {
+	
+	private List<DeveloperDecertifiedDTO> developerDecertificationResult;
+	
+	public DeveloperDecertificationResponse(){}
+
+	public List<DeveloperDecertifiedDTO> getDeveloperDecertificationResult() {
+		return developerDecertificationResult;
+	}
+
+	public void setDeveloperDecertificationResult(List<DeveloperDecertifiedDTO> developerDecertificationResult) {
+		this.developerDecertificationResult = developerDecertificationResult;
+	};
+
+}