+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/web/controller/SearchViewController.java
@@ -4,6 +4,8 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.OutputStream;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Set;
 
 import javax.servlet.http.HttpServletRequest;
@@ -25,14 +27,17 @@ import org.springframework.web.bind.annotation.RestController;
 
 import gov.healthit.chpl.dao.EntityRetrievalException;
 import gov.healthit.chpl.domain.CertifiedProductSearchDetails;
+import gov.healthit.chpl.domain.DeveloperDecertificationResponse;
 import gov.healthit.chpl.domain.KeyValueModel;
 import gov.healthit.chpl.domain.KeyValueModelStatuses;
 import gov.healthit.chpl.domain.PopulateSearchOptions;
 import gov.healthit.chpl.domain.SearchOption;
 import gov.healthit.chpl.domain.SearchRequest;
 import gov.healthit.chpl.domain.SearchResponse;
+import gov.healthit.chpl.dto.DeveloperDecertifiedDTO;
 import gov.healthit.chpl.manager.CertifiedProductDetailsManager;
 import gov.healthit.chpl.manager.CertifiedProductSearchManager;
+import gov.healthit.chpl.manager.DeveloperManager;
 import gov.healthit.chpl.manager.SearchMenuManager;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -52,6 +57,8 @@ public class SearchViewController {
 	@Autowired
 	private CertifiedProductDetailsManager certifiedProductDetailsManager;
 	
+	@Autowired
+	private DeveloperManager developerManager;
 	
 	private static final Logger logger = LogManager.getLogger(SearchViewController.class);
 
@@ -391,6 +398,22 @@ public class SearchViewController {
 			simple = false;
 		}
 		return searchMenuManager.getPopulateSearchOptions(simple);
+	}
+	
+	@ApiOperation(value="Get all developer decertifications in the CHPL", 
+			notes="This returns all decertified developers.")
+	@RequestMapping(value="/decertifications/developers", method=RequestMethod.GET,
+			produces="application/json; charset=utf-8")
+	public @ResponseBody DeveloperDecertificationResponse getDeveloperDecertifications() throws EntityRetrievalException {
+		DeveloperDecertificationResponse ddr = new DeveloperDecertificationResponse();
+		
+		List<DeveloperDecertifiedDTO> dtoList = new ArrayList<DeveloperDecertifiedDTO>();
+		
+		dtoList = developerManager.getDecertifiedDevelopers();
+		
+		ddr.setDeveloperDecertificationResult(dtoList);
+		
+		return ddr;
 	}
 	
 }