+++ chpl/chpl-service/src/test/java/gov/healthit/chpl/manager/impl/CertifiedProductSearchManagerTest.java
@@ -43,7 +43,7 @@ public class CertifiedProductSearchManagerTest extends TestCase {
 		SearchRequest searchRequest = new SearchRequest();
 		searchRequest.setDeveloper("Test Developer 1");
 		SearchResponse response = certifiedProductSearchManager.search(searchRequest);
-		assertEquals(3, response.getResults().size());
+		assertEquals(5, response.getResults().size());
 		
 		for (CertifiedProductSearchResult result : response.getResults() ){
 			assertTrue(result.getDeveloper().get("name").toString().startsWith("Test Developer 1"));
@@ -71,7 +71,7 @@ public class CertifiedProductSearchManagerTest extends TestCase {
 		SearchRequest searchRequest = new SearchRequest();
 		searchRequest.setVersion("1.0.1");
 		SearchResponse response = certifiedProductSearchManager.search(searchRequest);
-		assertEquals(1, response.getResults().size());
+		assertEquals(3, response.getResults().size());
 		
 		for (CertifiedProductSearchResult result : response.getResults() ){
 			assertTrue(result.getProduct().get("version").toString().startsWith("1.0.1"));
@@ -127,7 +127,7 @@ public class CertifiedProductSearchManagerTest extends TestCase {
 		
 		SearchRequest searchRequest = new SearchRequest();
 		SearchResponse response = certifiedProductSearchManager.search(searchRequest);
-		assertEquals(4, response.getResults().size());
+		assertEquals(12, response.getResults().size());
 	}
 	
 	@Test